alsa/hda/usb/core: Fix 'const' warnings and some retyping
Patchwork-id: 105308
O-Subject: [RHEL 6.7 PATCH 841/846] ALSA/HDA/USB/core: Fix 'const' warnings and some retyping
Bugzilla: 1069236
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

rhbz#1069236

+++ sound/pci/lx6464es/lx6464es.c
@@ -453,7 +453,7 @@ static void lx_trigger_start(struct lx6464es *chip, struct lx_stream *lx_stream)
 				     &buffer_index);
 
 		snd_printdd(LXP "starting: buffer index %x on %p (%d bytes)\n",
-			    buffer_index, (void *)buf, period_bytes);
+			    buffer_index, (void *)(uintptr_t)buf, period_bytes);
 		buf += period_bytes;
 	}
